This commit introduces two major new features: "Create Shortcut" and a file association framework for "Open With", and fixes related bugs.
**Create Shortcut:**
- Users can now right-click on any application in the Start Menu and select "Create shortcut".
- This action creates a `.app` file on the Desktop, which acts as a shortcut to the original application.

**"Open With" Framework:**
- A new `fileAssociationService` has been created to manage which applications can open specific file extensions.
- Applications can now declare the file types they support in their definition.
- The file context menu now includes an "Open with" sub-menu, dynamically showing all compatible applications for the selected file.
- Double-click logic has been refactored to use this service to open files with their default associated app.

**Bug Fixes:**
- Corrected an issue where Start Menu context menu actions were not firing.
- Fixed a caching bug in `getAppDefinitions` that prevented the "Open With" menu from populating correctly.
